extract processors
