+++ packages/ddp-client/common/livedata_connection.js
@@ -15,6 +15,8 @@ import {
 } from "meteor/ddp-common/utils";
 import { ConnectionStreamHandlers } from './connection_stream_handlers';
 import { MongoIDMap } from './mongo_id_map';
+import { MessageProcessors } from './message_processors';
+import { DocumentProcessors } from './document_processors';
 
 // @param url {String|Object} URL to Meteor app,
 //   or an object as a test hook (see code)
@@ -272,6 +274,30 @@ export class Connection {
       this._stream.on('reset', () => this._streamHandlers.onReset());
       this._stream.on('disconnect', onDisconnect);
     }
+
+    this._messageProcessors = new MessageProcessors(this);
+
+    // Expose message processor methods to maintain backward compatibility
+    this._livedata_connected = (msg) => this._messageProcessors._livedata_connected(msg);
+    this._livedata_data = (msg) => this._messageProcessors._livedata_data(msg);
+    this._livedata_nosub = (msg) => this._messageProcessors._livedata_nosub(msg);
+    this._livedata_result = (msg) => this._messageProcessors._livedata_result(msg);
+    this._livedata_error = (msg) => this._messageProcessors._livedata_error(msg);
+
+    this._documentProcessors = new DocumentProcessors(this);
+
+    // Expose document processor methods to maintain backward compatibility
+    this._process_added = (msg, updates) => this._documentProcessors._process_added(msg, updates);
+    this._process_changed = (msg, updates) => this._documentProcessors._process_changed(msg, updates);
+    this._process_removed = (msg, updates) => this._documentProcessors._process_removed(msg, updates);
+    this._process_ready = (msg, updates) => this._documentProcessors._process_ready(msg, updates);
+    this._process_updated = (msg, updates) => this._documentProcessors._process_updated(msg, updates);
+
+    // Also expose utility methods used by other parts of the system
+    this._pushUpdate = (updates, collection, msg) =>
+      this._documentProcessors._pushUpdate(updates, collection, msg);
+    this._getServerDoc = (collection, id) =>
+      this._documentProcessors._getServerDoc(collection, id);
   }
 
   // 'name' is the name of the data on the wire that should go in the
@@ -1094,121 +1120,6 @@ export class Connection {
     return Object.values(invokers).some((invoker) => !!invoker.sentMessage);
   }
 
-  async _livedata_connected(msg) {
-    const self = this;
-
-    if (self._version !== 'pre1' && self._heartbeatInterval !== 0) {
-      self._heartbeat = new DDPCommon.Heartbeat({
-        heartbeatInterval: self._heartbeatInterval,
-        heartbeatTimeout: self._heartbeatTimeout,
-        onTimeout() {
-          self._lostConnection(
-            new DDP.ConnectionError('DDP heartbeat timed out')
-          );
-        },
-        sendPing() {
-          self._send({ msg: 'ping' });
-        }
-      });
-      self._heartbeat.start();
-    }
-
-    // If this is a reconnect, we'll have to reset all stores.
-    if (self._lastSessionId) self._resetStores = true;
-
-    let reconnectedToPreviousSession;
-    if (typeof msg.session === 'string') {
-      reconnectedToPreviousSession = self._lastSessionId === msg.session;
-      self._lastSessionId = msg.session;
-    }
-
-    if (reconnectedToPreviousSession) {
-      // Successful reconnection -- pick up where we left off.  Note that right
-      // now, this never happens: the server never connects us to a previous
-      // session, because DDP doesn't provide enough data for the server to know
-      // what messages the client has processed. We need to improve DDP to make
-      // this possible, at which point we'll probably need more code here.
-      return;
-    }
-
-    // Server doesn't have our data any more. Re-sync a new session.
-
-    // Forget about messages we were buffering for unknown collections. They'll
-    // be resent if still relevant.
-    self._updatesForUnknownStores = Object.create(null);
-
-    if (self._resetStores) {
-      // Forget about the effects of stubs. We'll be resetting all collections
-      // anyway.
-      self._documentsWrittenByStub = Object.create(null);
-      self._serverDocuments = Object.create(null);
-    }
-
-    // Clear _afterUpdateCallbacks.
-    self._afterUpdateCallbacks = [];
-
-    // Mark all named subscriptions which are ready (ie, we already called the
-    // ready callback) as needing to be revived.
-    // XXX We should also block reconnect quiescence until unnamed subscriptions
-    //     (eg, autopublish) are done re-publishing to avoid flicker!
-    self._subsBeingRevived = Object.create(null);
-    Object.entries(self._subscriptions).forEach(([id, sub]) => {
-      if (sub.ready) {
-        self._subsBeingRevived[id] = true;
-      }
-    });
-
-    // Arrange for "half-finished" methods to have their callbacks run, and
-    // track methods that were sent on this connection so that we don't
-    // quiesce until they are all done.
-    //
-    // Start by clearing _methodsBlockingQuiescence: methods sent before
-    // reconnect don't matter, and any "wait" methods sent on the new connection
-    // that we drop here will be restored by the loop below.
-    self._methodsBlockingQuiescence = Object.create(null);
-    if (self._resetStores) {
-      const invokers = self._methodInvokers;
-      keys(invokers).forEach(id => {
-        const invoker = invokers[id];
-        if (invoker.gotResult()) {
-          // This method already got its result, but it didn't call its callback
-          // because its data didn't become visible. We did not resend the
-          // method RPC. We'll call its callback when we get a full quiesce,
-          // since that's as close as we'll get to "data must be visible".
-          self._afterUpdateCallbacks.push(
-            (...args) => invoker.dataVisible(...args)
-          );
-        } else if (invoker.sentMessage) {
-          // This method has been sent on this connection (maybe as a resend
-          // from the last connection, maybe from onReconnect, maybe just very
-          // quickly before processing the connected message).
-          //
-          // We don't need to do anything special to ensure its callbacks get
-          // called, but we'll count it as a method which is preventing
-          // reconnect quiescence. (eg, it might be a login method that was run
-          // from onReconnect, and we don't want to see flicker by seeing a
-          // logged-out state.)
-          self._methodsBlockingQuiescence[invoker.methodId] = true;
-        }
-      });
-    }
-
-    self._messagesBufferedUntilQuiescence = [];
-
-    // If we're not waiting on any methods or subs, we can reset the stores and
-    // call the callbacks immediately.
-    if (! self._waitingForQuiescence()) {
-      if (self._resetStores) {
-        for (const store of Object.values(self._stores)) {
-          await store.beginUpdate(0, true);
-          await store.endUpdate();
-        }
-        self._resetStores = false;
-      }
-      self._runAfterUpdateCallbacks();
-    }
-  }
-
   async _processOneDataMessage(msg, updates) {
     const messageType = msg.msg;
 
@@ -1228,87 +1139,6 @@ export class Connection {
     } else {
       Meteor._debug('discarding unknown livedata data message type', msg);
     }
-  }
-
-  async _livedata_data(msg) {
-    const self = this;
-
-    if (self._waitingForQuiescence()) {
-      self._messagesBufferedUntilQuiescence.push(msg);
-
-      if (msg.msg === 'nosub') {
-        delete self._subsBeingRevived[msg.id];
-      }
-
-      if (msg.subs) {
-        msg.subs.forEach(subId => {
-          delete self._subsBeingRevived[subId];
-        });
-      }
-
-      if (msg.methods) {
-        msg.methods.forEach(methodId => {
-          delete self._methodsBlockingQuiescence[methodId];
-        });
-      }
-
-      if (self._waitingForQuiescence()) {
-        return;
-      }
-
-      // No methods or subs are blocking quiescence!
-      // We'll now process and all of our buffered messages, reset all stores,
-      // and apply them all at once.
-
-      const bufferedMessages = self._messagesBufferedUntilQuiescence;
-      for (const bufferedMessage of Object.values(bufferedMessages)) {
-        await self._processOneDataMessage(
-          bufferedMessage,
-          self._bufferedWrites
-        );
-      }
-
-      self._messagesBufferedUntilQuiescence = [];
-
-    } else {
-      await self._processOneDataMessage(msg, self._bufferedWrites);
-    }
-
-    // Immediately flush writes when:
-    //  1. Buffering is disabled. Or;
-    //  2. any non-(added/changed/removed) message arrives.
-    const standardWrite =
-      msg.msg === "added" ||
-      msg.msg === "changed" ||
-      msg.msg === "removed";
-
-    if (self._bufferedWritesInterval === 0 || ! standardWrite) {
-      await self._flushBufferedWrites();
-      return;
-    }
-
-    if (self._bufferedWritesFlushAt === null) {
-      self._bufferedWritesFlushAt =
-        new Date().valueOf() + self._bufferedWritesMaxAge;
-    } else if (self._bufferedWritesFlushAt < new Date().valueOf()) {
-      await self._flushBufferedWrites();
-      return;
-    }
-
-    if (self._bufferedWritesFlushHandle) {
-      clearTimeout(self._bufferedWritesFlushHandle);
-    }
-    self._bufferedWritesFlushHandle = setTimeout(() => {
-      // __flushBufferedWrites is a promise, so with this we can wait the promise to finish
-      // before doing something
-      self._liveDataWritesPromise = self._flushBufferedWrites();
-
-      if (Meteor._isPromise(self._liveDataWritesPromise)) {
-        self._liveDataWritesPromise.finally(
-          () => (self._liveDataWritesPromise = undefined)
-        );
-      }
-    }, self._bufferedWritesInterval);
   }
 
   _prepareBuffersToFlush() {
@@ -1438,160 +1268,6 @@ export class Connection {
     });
   }
 
-  _pushUpdate(updates, collection, msg) {
-    if (! hasOwn.call(updates, collection)) {
-      updates[collection] = [];
-    }
-    updates[collection].push(msg);
-  }
-
-  _getServerDoc(collection, id) {
-    const self = this;
-    if (! hasOwn.call(self._serverDocuments, collection)) {
-      return null;
-    }
-    const serverDocsForCollection = self._serverDocuments[collection];
-    return serverDocsForCollection.get(id) || null;
-  }
-
-  async _process_added(msg, updates) {
-    const self = this;
-    const id = MongoID.idParse(msg.id);
-    const serverDoc = self._getServerDoc(msg.collection, id);
-    if (serverDoc) {
-      // Some outstanding stub wrote here.
-      const isExisting = serverDoc.document !== undefined;
-
-      serverDoc.document = msg.fields || Object.create(null);
-      serverDoc.document._id = id;
-
-      if (self._resetStores) {
-        // During reconnect the server is sending adds for existing ids.
-        // Always push an update so that document stays in the store after
-        // reset. Use current version of the document for this update, so
-        // that stub-written values are preserved.
-        const currentDoc = await self._stores[msg.collection].getDoc(msg.id);
-        if (currentDoc !== undefined) msg.fields = currentDoc;
-
-        self._pushUpdate(updates, msg.collection, msg);
-      } else if (isExisting) {
-        throw new Error('Server sent add for existing id: ' + msg.id);
-      }
-    } else {
-      self._pushUpdate(updates, msg.collection, msg);
-    }
-  }
-
-  _process_changed(msg, updates) {
-    const self = this;
-    const serverDoc = self._getServerDoc(msg.collection, MongoID.idParse(msg.id));
-    if (serverDoc) {
-      if (serverDoc.document === undefined)
-        throw new Error('Server sent changed for nonexisting id: ' + msg.id);
-      DiffSequence.applyChanges(serverDoc.document, msg.fields);
-    } else {
-      self._pushUpdate(updates, msg.collection, msg);
-    }
-  }
-
-  _process_removed(msg, updates) {
-    const self = this;
-    const serverDoc = self._getServerDoc(msg.collection, MongoID.idParse(msg.id));
-    if (serverDoc) {
-      // Some outstanding stub wrote here.
-      if (serverDoc.document === undefined)
-        throw new Error('Server sent removed for nonexisting id:' + msg.id);
-      serverDoc.document = undefined;
-    } else {
-      self._pushUpdate(updates, msg.collection, {
-        msg: 'removed',
-        collection: msg.collection,
-        id: msg.id
-      });
-    }
-  }
-
-  _process_updated(msg, updates) {
-    const self = this;
-    // Process "method done" messages.
-
-    msg.methods.forEach((methodId) => {
-      const docs = self._documentsWrittenByStub[methodId] || {};
-      Object.values(docs).forEach((written) => {
-        const serverDoc = self._getServerDoc(written.collection, written.id);
-        if (! serverDoc) {
-          throw new Error('Lost serverDoc for ' + JSON.stringify(written));
-        }
-        if (! serverDoc.writtenByStubs[methodId]) {
-          throw new Error(
-            'Doc ' +
-              JSON.stringify(written) +
-              ' not written by  method ' +
-              methodId
-          );
-        }
-        delete serverDoc.writtenByStubs[methodId];
-        if (isEmpty(serverDoc.writtenByStubs)) {
-          // All methods whose stubs wrote this method have completed! We can
-          // now copy the saved document to the database (reverting the stub's
-          // change if the server did not write to this object, or applying the
-          // server's writes if it did).
-
-          // This is a fake ddp 'replace' message.  It's just for talking
-          // between livedata connections and minimongo.  (We have to stringify
-          // the ID because it's supposed to look like a wire message.)
-          self._pushUpdate(updates, written.collection, {
-            msg: 'replace',
-            id: MongoID.idStringify(written.id),
-            replace: serverDoc.document
-          });
-          // Call all flush callbacks.
-
-          serverDoc.flushCallbacks.forEach((c) => {
-            c();
-          });
-
-          // Delete this completed serverDocument. Don't bother to GC empty
-          // IdMaps inside self._serverDocuments, since there probably aren't
-          // many collections and they'll be written repeatedly.
-          self._serverDocuments[written.collection].remove(written.id);
-        }
-      });
-      delete self._documentsWrittenByStub[methodId];
-
-      // We want to call the data-written callback, but we can't do so until all
-      // currently buffered messages are flushed.
-      const callbackInvoker = self._methodInvokers[methodId];
-      if (! callbackInvoker) {
-        throw new Error('No callback invoker for method ' + methodId);
-      }
-
-      self._runWhenAllServerDocsAreFlushed(
-        (...args) => callbackInvoker.dataVisible(...args)
-      );
-    });
-  }
-
-  _process_ready(msg, updates) {
-    const self = this;
-    // Process "sub ready" messages. "sub ready" messages don't take effect
-    // until all current server documents have been flushed to the local
-    // database. We can use a write fence to implement this.
-
-    msg.subs.forEach((subId) => {
-      self._runWhenAllServerDocsAreFlushed(() => {
-        const subRecord = self._subscriptions[subId];
-        // Did we already unsubscribe?
-        if (!subRecord) return;
-        // Did we already receive a ready message? (Oops!)
-        if (subRecord.ready) return;
-        subRecord.ready = true;
-        subRecord.readyCallback && subRecord.readyCallback();
-        subRecord.readyDeps.changed();
-      });
-    });
-  }
-
   // Ensures that "f" will be called after all documents currently in
   // _serverDocuments have been written to the local cache. f will not be called
   // if the connection is lost before then!
@@ -1628,93 +1304,6 @@ export class Connection {
       // There aren't any buffered docs --- we can call f as soon as the current
       // round of updates is applied!
       runFAfterUpdates();
-    }
-  }
-
-  async _livedata_nosub(msg) {
-    const self = this;
-
-    // First pass it through _livedata_data, which only uses it to help get
-    // towards quiescence.
-    await self._livedata_data(msg);
-
-    // Do the rest of our processing immediately, with no
-    // buffering-until-quiescence.
-
-    // we weren't subbed anyway, or we initiated the unsub.
-    if (! hasOwn.call(self._subscriptions, msg.id)) {
-      return;
-    }
-
-    // XXX COMPAT WITH 1.0.3.1 #errorCallback
-    const errorCallback = self._subscriptions[msg.id].errorCallback;
-    const stopCallback = self._subscriptions[msg.id].stopCallback;
-
-    self._subscriptions[msg.id].remove();
-
-    const meteorErrorFromMsg = msgArg => {
-      return (
-        msgArg &&
-        msgArg.error &&
-        new Meteor.Error(
-          msgArg.error.error,
-          msgArg.error.reason,
-          msgArg.error.details
-        )
-      );
-    };
-
-    // XXX COMPAT WITH 1.0.3.1 #errorCallback
-    if (errorCallback && msg.error) {
-      errorCallback(meteorErrorFromMsg(msg));
-    }
-
-    if (stopCallback) {
-      stopCallback(meteorErrorFromMsg(msg));
-    }
-  }
-
-  async _livedata_result(msg) {
-    // id, result or error. error has error (code), reason, details
-
-    const self = this;
-
-    // Lets make sure there are no buffered writes before returning result.
-    if (! isEmpty(self._bufferedWrites)) {
-      await self._flushBufferedWrites();
-    }
-
-    // find the outstanding request
-    // should be O(1) in nearly all realistic use cases
-    if (isEmpty(self._outstandingMethodBlocks)) {
-      Meteor._debug('Received method result but no methods outstanding');
-      return;
-    }
-    const currentMethodBlock = self._outstandingMethodBlocks[0].methods;
-    let i;
-    const m = currentMethodBlock.find((method, idx) => {
-      const found = method.methodId === msg.id;
-      if (found) i = idx;
-      return found;
-    });
-    if (!m) {
-      Meteor._debug("Can't match method response to original method call", msg);
-      return;
-    }
-
-    // Remove from current method block. This may leave the block empty, but we
-    // don't move on to the next block until the callback has been delivered, in
-    // _outstandingMethodFinished.
-    currentMethodBlock.splice(i, 1);
-
-    if (hasOwn.call(msg, 'error')) {
-      m.receiveResult(
-        new Meteor.Error(msg.error.error, msg.error.reason, msg.error.details)
-      );
-    } else {
-      // msg.result may be undefined if the method didn't return a
-      // value
-      m.receiveResult(undefined, msg.result);
     }
   }
 
@@ -1784,11 +1373,6 @@ export class Connection {
     self._outstandingMethodBlocks[0].methods.forEach(m => {
       m.sendMessage();
     });
-  }
-
-  _livedata_error(msg) {
-    Meteor._debug('Received error from server: ', msg.reason);
-    if (msg.offendingMessage) Meteor._debug('For: ', msg.offendingMessage);
   }
 
   _sendOutstandingMethodBlocksMessages(oldOutstandingMethodBlocks) {
